Added tests the following tests in the file test_harvester_global_soilt4.py test_gridcell_area_conservation() test_global_mean_values() test_global_mean_values2() test_gridcell_variance() test_gridcell_min_max() Removed bfg files for prateb in test/data.  Since these are not used in this test.
